debundle/lowering: extract chunk_ast
BAZEL_TEST_INVOCATIONS=none: file move.
